Rename the proxy
