+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/extractionservice/ExtractionResource.java
@@ -9,7 +9,7 @@ import software.uncharted.terarium.hmiserver.exceptions.HmiResponseExceptionMapp
 import software.uncharted.terarium.hmiserver.models.dataservice.Artifact;
 import software.uncharted.terarium.hmiserver.models.dataservice.Model;
 import software.uncharted.terarium.hmiserver.models.extractionservice.ExtractionResponse;
-import software.uncharted.terarium.hmiserver.proxies.extractionservice.ExtractionServiceProxy;
+import software.uncharted.terarium.hmiserver.proxies.knowledgemiddleware.KnowledgeMiddlewareProxy;
 import software.uncharted.terarium.hmiserver.proxies.skema.SkemaUnifiedProxy;
 import software.uncharted.terarium.hmiserver.proxies.dataservice.ArtifactProxy;
 
@@ -27,7 +27,7 @@ public class ExtractionResource {
 
 	@Inject
 	@RestClient
-	ExtractionServiceProxy extractionProxy;
+	KnowledgeMiddlewareProxy knowledgeMiddlewareProxy;
 
 	@Inject
 	@RestClient
@@ -46,7 +46,7 @@ public class ExtractionResource {
 	@Path("/status/{id}")
 	public Response getTaskStatus(
 		@PathParam("id") final String id) {
-		return extractionProxy.getTaskStatus(id);
+		return knowledgeMiddlewareProxy.getTaskStatus(id);
 	}
 
 	/**
@@ -68,7 +68,7 @@ public class ExtractionResource {
 		@DefaultValue("petrinet") @QueryParam("framework") String framework,
 		List<String> mathMLPayload
 	) {
-		return extractionProxy.postMathMLToAMR(framework, mathMLPayload);
+		return knowledgeMiddlewareProxy.postMathMLToAMR(framework, mathMLPayload);
 	};
 
 	/**
@@ -115,7 +115,7 @@ public class ExtractionResource {
 		if (name == null) {	name = artifact.getName(); }
 		if (description == null) { description = artifact.getDescription();	}
 
-		return extractionProxy.postCodeToAMR(artifactId, name, description);
+		return knowledgeMiddlewareProxy.postCodeToAMR(artifactId, name, description);
 	}
 
 
@@ -142,7 +142,7 @@ public class ExtractionResource {
 		@QueryParam("name") String name,
 		@QueryParam("description") String description
 	) {
-		return extractionProxy.postPDFExtractions(artifactId, annotateSkema, annotateMIT, name, description);
+		return knowledgeMiddlewareProxy.postPDFExtractions(artifactId, annotateSkema, annotateMIT, name, description);
 	};
 
 	/**
@@ -153,7 +153,7 @@ public class ExtractionResource {
 	@POST
 	@Path("/pdf-to-text")
 	public Response postPDFToText(@QueryParam("artifact_id") String artifactId){
-		return extractionProxy.postPDFToText(artifactId);
+		return knowledgeMiddlewareProxy.postPDFToText(artifactId);
 	}
 
 	/**
@@ -171,6 +171,6 @@ public class ExtractionResource {
 		@PathParam("dataset_id") String datasetId,
 		@QueryParam("artifact_id") String artifactId
 	) {
-		return extractionProxy.postProfileDataset(datasetId, artifactId);
+		return knowledgeMiddlewareProxy.postProfileDataset(datasetId, artifactId);
 	};
 }
